third_party/WebKit/LayoutTests/editing/inserting/insert_html_to_replace_selection.html

<!doctype html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../assert_selection.js"></script>
<div id="log"></div>
<script>
test(() => assert_selection(
    '<div contenteditable>A new ^link|</div>',
    'insertHTML <a href="http://www.google.com">link</a>',
    '<div contenteditable>A new\u{00A0}<a href="http://www.google.com">link|</a></div>'),
    'Replace selection by insertHTML');
</script>
